Refactoring$
